+++ components/button/Button.tsx
@@ -4,11 +4,13 @@ import { ConfigContext } from '../provider/context'
 import './style/Button.scss'
 import components from '../provider/components'
 
-export type ButtonCap = 'default' | 'primary' | 'ghost' | 'dashed' | 'danger' | 'font';
+export type ButtonCap = 'default' | 'primary' | 'text' | 'ghost' | 'dashed' | 'danger';
+export type ButtonSize = 'large' | 'middle' | 'small';
 
 export interface BaseButtonProps {
   cap?: ButtonCap;
   label?: string;
+  size?: ButtonSize;
 }
 
 export type ButtonProps = BaseButtonProps & React.ButtonHTMLAttributes<HTMLButtonElement>;
@@ -18,16 +20,21 @@ const InternalButton: React.ForwardRefRenderFunction<unknown, ButtonProps> = (pr
     className,
     children,
     cap,
-    label
+    label,
+    size
   } = props
 
   const buttonRef = (ref as any) || React.createRef<HTMLElement>()
 
   const { getPrefixCls } = useContext(ConfigContext)
   const prefixCls = getPrefixCls(components.BUTTON)
-  const buttonCls = classNames(prefixCls, className, {
-    [`${prefixCls}-${cap}`]: cap
-  })
+  const buttonCls = classNames(
+    prefixCls,
+    {
+      [`${prefixCls}-${cap}`]: cap,
+      [`${prefixCls}-${size}`]: size
+    },
+    className)
 
   const buttonNode = label == null
     ? (
@@ -49,7 +56,8 @@ const Button = React.forwardRef<unknown, ButtonProps>(InternalButton)
 Button.displayName = 'Button'
 
 Button.defaultProps = {
-  cap: 'default'
+  cap: 'default',
+  size: 'middle'
 }
 
 export default Button