+++ components/index.tsx
@@ -1,11 +1,17 @@
 export { default as Button } from './button/Button'
 
+export { default as Upload } from './upload/Upload'
+
 export { default as Modal } from './modal/Modal'
 
-export { default as Input } from './input/input'
+export { default as Input } from './input/Input'
 
 export { default as Alert } from './alert/alert'
+
+// export { default as Dropdowns } from './dropdowns/Dropdown'
 
 export { default as Message } from './message/message'
 
 export { Utils } from './common/utils'
+
+export { keycap, RequestMethod } from './axios/index'